Fixed bug
Fixed a bug the program had with a win scenario on the 9th move. It now checks if no win has happened by the 9th move before declaring a draw.
Additionally, I removed some div elements that were not necessary as well as some slight CSS changes to make sure the buttons are responsive.